Make it possible to address cross-origin windows with testdriver.
The current testdriver setup used a property __wptrunner_id on the
window to store a relationship between js window objects and webdriver
windows; the property is set in the window invoking the testdriver
command (typically the test window, but can also be another window as
long as it can get a handle to the test window).

The problem with this setup is it doesn't work well for cross-origin
cases, particularly in cases where it isn't possible to get a js
object representing the remote window in the test window, or
vice-versa (e.g. window.open with rel=noopener).

The ideal solution here would be to have the WebDriver layer provide
an id for each window, but that currently isn't implemented. Instead
we provide some more options to make it possible to set up
cross-window communication:

UUIDs may be provided on a URL, by adding a query parameter with name
'uuid'. When looking up these UUIDs we look both at the src of the
element containing a nested browsing context, and the location of the
URL itself. The advantage of the former is that we can often access
the container same-origin in cases where the nested context is
cross-origin, and the UUID remains constant even if the browsing
context undergoes navigation.

On the testdriver side, the functions that take a context no longer
need to be passed a js window handle, but can also take a containing
element or an explicit UUID that's communicated out of band
somehow (e.g. for a noreferrer top-level browsing context where the
URL is constructed with a known uuid parameter).

This is obviously pretty complex, but it's hard to see how to do
better without some additional support on the WebDriver side.