Refactor XR mojo API to directly send array of views
This change sends the array of views directly from the XR process to
Blink and the browser process, instead of sending individual
head-from-eye and head-from-local poses. With this change, Blink
and the browser process no longer have to manually calculate the
transforms and can instead use the transforms directly from the XR
runtime. This also allows the XR runtime to send an arbitrary number
of views in the case where WebXR requests secondary views.

This change also improves OpenXR stability. Requesting head-from-eye
and head-from-local poses requires two separate calls to OpenXR,
which uses slightly different timestamps. As a result, the final
transform calculated by Blink and passed to WebXR is slightly off
from the transform OpenXR use, resulting in some jittering in
certain devices.

Change-Id: I5da254537ce1317e43677863aeeb6c0f234957f8
Bug: 998146